Compress json_token from 48 bytes to 16
Since the library primarily deals with allocating and populating arrays
of json_tokens, it makes sense to try to keep their size reasonable.
Unfortunately, I didn't do that when I initially wrote the library. I
believed that the only future proof way to write software was with
64-bit values ("it's a size_t, that's tho correct type!"). I also didn't
stop to think what values were important.

So the json_token had an "end" field, despite that you can look at the
next token's "start" field to get a good idea of it. Similarly, the
"child" field existed, despite the fact that you could always compute
the child index directly! So embarrassing.

By removing these two useless fields, and by shrinking the integer
fields to a (still respectable) unsigned 32-bit int, the size is reduced
by 67%.

+++ src/json_private.h
@@ -29,8 +29,8 @@ extern char *json_type_str[JSON_NULL + 1];
 extern char *json_error_str[_LAST_JSONERR];
 
 void json_settoken(struct json_token *arr, struct json_token tok,
-                   struct json_parser p, size_t maxtoken);
+                   struct json_parser p, uint32_t maxtoken);
 struct json_parser json_parse_string(const char *text, struct json_token *arr,
-                                     size_t maxtoken, struct json_parser p);
+                                     uint32_t maxtoken, struct json_parser p);
 
 #endif // SMB_JSON_PRIVATE_H
