+++ test/parse_numbers.c
@@ -30,7 +30,7 @@ void tearDown(void)
 static void test_single_digit(void)
 {
 	char input[] = "0";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -44,7 +44,7 @@ static void test_single_digit(void)
 static void test_multiple_digit(void)
 {
 	char input[] = "12";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -76,7 +76,7 @@ static void test_starts_with_zero(void)
 static void test_decimal(void)
 {
 	char input[] = "1.1";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -90,7 +90,7 @@ static void test_decimal(void)
 static void test_zero_decimal(void)
 {
 	char input[] = "0.1";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -104,7 +104,7 @@ static void test_zero_decimal(void)
 static void test_negative_sign(void)
 {
 	char input[] = "-1";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -118,7 +118,7 @@ static void test_negative_sign(void)
 static void test_exponent_upper(void)
 {
 	char input[] = "1E5";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -132,7 +132,7 @@ static void test_exponent_upper(void)
 static void test_exponent_lower(void)
 {
 	char input[] = "1e5";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -146,7 +146,7 @@ static void test_exponent_lower(void)
 static void test_exponent_plus(void)
 {
 	char input[] = "1e+5";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -160,7 +160,7 @@ static void test_exponent_plus(void)
 static void test_exponent_minus(void)
 {
 	char input[] = "1e-5";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -174,7 +174,7 @@ static void test_exponent_minus(void)
 static void test_sign_decimal_exponent(void)
 {
 	char input[] = "-1.5e+5";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -216,7 +216,7 @@ static void test_exponent_sign_without_digits(void)
 static void test_negative_zero(void)
 {
 	char input[] = "-0"; // believe it or not, this is valid JSON.
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -230,7 +230,7 @@ static void test_negative_zero(void)
 static void test_zero_exp(void)
 {
 	char input[] = "0e5"; // again, doesn't make sense, but is valid
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -244,7 +244,7 @@ static void test_zero_exp(void)
 static void test_double_digit_decimal(void)
 {
 	char input[] = "1.23";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number;
@@ -258,7 +258,7 @@ static void test_double_digit_decimal(void)
 static void test_double_digit_exp(void)
 {
 	char input[] = "1e23";
-	size_t ntok = 1;
+	uint32_t ntok = 1;
 	struct json_token tokens[ntok];
 	struct json_parser p = json_parse(input, tokens, ntok);
 	double number = 0;
